requirements.txt
pyjwt
pydantic
cryptography
sqlalchemy
pymysql
python-dotenv